requirements.txt
# deployment
ansible

# smoke testing
python-keystoneclient
requests

# load testing
dataset
locustio
pyzmq # for distributed locust
